Add initial documentation
Add an initial user guide, server REST API documentation, and client
API documentation. Add the `docs` session to the nox configuration
to build the documentation.

Split the client exceptions into a separate file instead of bundling
them with the client.
