+++ .pre-commit-config.yaml
@@ -12,7 +12,7 @@ repos:
       - id: uv-lock
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.12.11
+    rev: v0.12.12
     hooks:
       - id: ruff-check
         args: [--fix, --exit-non-zero-on-fix]
