+++ client/src/rubin/repertoire/__init__.py
@@ -1,27 +1,37 @@
 """Client, models, and URL construction for Repertoire."""
 
 from ._builder import RepertoireBuilder
-from ._client import (
-    DiscoveryClient,
+from ._client import DiscoveryClient
+from ._config import (
+    BaseRule,
+    DataServiceRule,
+    DatasetConfig,
+    InternalServiceRule,
+    RepertoireSettings,
+    UiServiceRule,
+)
+from ._exceptions import (
     RepertoireError,
     RepertoireUrlError,
     RepertoireValidationError,
     RepertoireWebError,
 )
-from ._config import DatasetConfig, RepertoireSettings
 from ._models import Dataset, Discovery, ServiceUrls
 
 __all__ = [
+    "BaseRule",
+    "DataServiceRule",
     "Dataset",
     "DatasetConfig",
     "Discovery",
     "DiscoveryClient",
+    "InternalServiceRule",
     "RepertoireBuilder",
     "RepertoireError",
     "RepertoireSettings",
     "RepertoireUrlError",
     "RepertoireValidationError",
     "RepertoireWebError",
-    "Rule",
     "ServiceUrls",
+    "UiServiceRule",
 ]
