+++ docs/dev/index.rst
@@ -1,10 +1,10 @@
-:og:description: Learn how to contribute to the repertoire open source project.
+:og:description: Learn how to contribute to the Repertoire open source project.
 
 ############
 Contributing
 ############
 
-Learn how to contribute to the repertoire open source project.
+Learn how to contribute to the Repertoire open source project.
 
 .. toctree::
    :caption: Guides
