+++ docs/dev/release.rst
@@ -2,14 +2,14 @@
 Release procedure
 #################
 
-This page gives an overview of how repertoire releases are made.
+This page gives an overview of how Repertoire releases are made.
 This information is only useful for maintainers.
 
-repertoire's releases are largely automated through GitHub Actions (see the `ci.yaml`_ workflow file for details).
-When a semantic version tag is pushed to GitHub, `repertoire is released to PyPI`_ with that version.
+Repertoire's releases are largely automated through GitHub Actions (see the `ci.yaml`_ workflow file for details).
+When a semantic version tag is pushed to GitHub, the Repertoire client is `released to PyPI`_ with that version.
 Similarly, documentation is built and pushed for each version (see https://repertoire.lsst.io/v).
 
-.. _`repertoire is released to PyPI`: https://pypi.org/project/repertoire/
+.. _`released to PyPI`: https://pypi.org/project/rubin-repertoire/
 .. _`ci.yaml`: https://github.com/lsst-sqre/repertoire/blob/main/.github/workflows/ci.yaml
 
 .. _regular-release:
@@ -33,7 +33,7 @@ When it comes time to make the release, there should be a collection of change l
 Those fragments will make up the change log for the new release.
 
 Review those fragments to determine the version number of the next release.
-Safir follows semver_, so follow its rules to pick the next version:
+Repertoire follows semver_, so follow its rules (summarized below) to pick the next version:
 
 .. rst-class:: compact
 
@@ -41,9 +41,10 @@ Safir follows semver_, so follow its rules to pick the next version:
 - If there are any new features, increment the minor version number and set the patch version to 0.
 - Otherwise, increment the patch version number.
 
-Then, run ``scriv collect --version <version>`` specifying the version number you decided on.
+Then, run :command:`uv run scriv collect --version <version>`, specifying the version number you decided on.
 This will delete the fragment files and collect them into :file:`CHANGELOG.md` under an entry for the new release.
 Review that entry and edit it as needed (proofread, change the order to put more important things first, etc.).
+
 scriv will put blank lines between entries from different files.
 You may wish to remove those blank lines to ensure consistent formatting by various Markdown parsers.
 
@@ -54,27 +55,29 @@ Finally, create a PR from those changes and merge it before continuing with the 
 
 Use `GitHub's Release feature <https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository>`__ to create releases and their corresponding Git tags.
 
-1. Specify a tag from the appropriate branch (typically ``main``).
-   This tag's name is :pep:`440` and is usually formatted at ``X.Y.Z`` (without a ``v`` prefix).
+1. For the tag, enter the version number of the release in the :guilabel:`Find or create a new tag` box in the dropdown under :guilabel:`Select tag`.
+   The tag must follow the :pep:`440` specification since Repertoire uses setuptools_scm_ to set version metadata based on Git tags.
+   In particular, don't prefix the tag with ``v``.
 
-2. For the release name, repeat the version string.
+   .. _setuptools_scm: https://github.com/pypa/setuptools_scm
 
-3. Fill in the release notes, copied from the release notes.
-   You can use GitHub's change log feature to additionally generate a list of PRs included in the release.
+2. Ensure the branch target is set appropriately (normally ``main``).
 
-The tag **must** follow the :pep:`440` specification since repertoire uses setuptools_scm_ to set version metadata based on Git tags.
-In particular, **don't** prefix the tag with ``v``.
+3. For the release title, repeat the version string.
 
-.. _setuptools_scm: https://github.com/pypa/setuptools_scm
+4. Click the :guilabel:`Generate release notes` button to include the GitHub-generated summary of pull requests in the release notes.
 
-The `ci.yaml`_ GitHub Actions workflow uploads the new release to PyPI and documentation to https://repertoire.lsst.io.
+5. In the release notes box above the generated notes, paste the contents of the :file:`CHANGELOG.md` entry for this release, without the initial heading specifying the version number and date.
+   Adjust the heading depth of the subsections to use ``##`` instead of ``###`` to match the pull request summary.
+
+The `ci.yaml`_ GitHub Actions workflow will upload the new release to PyPI, documentation to https://repertoire.lsst.io, and a Docker image to the GitHub Container Registry.
 
 .. _backport-release:
 
 Backport releases
 =================
 
-The regular release procedure works from the main line of development on the ``master`` Git branch.
+The regular release procedure works from the main line of development on the ``main`` Git branch.
 To create a release that patches an earlier major or minor version, you need to release from a **release branch.**
 
 Creating a release branch
@@ -86,7 +89,7 @@ If the release branch doesn't already exist, check out the latest patch for that
 .. code-block:: sh
 
    git checkout X.Y.Z
-   git checkout -b X.Y
+   git switch -c X.Y
    git push -u
 
 Developing on a release branch
