+++ docs/user-guide/index.rst
@@ -1,9 +1,27 @@
-:og:description: Learn how to use repertoire.
+:og:description: Learn how to use Repertoire.
 
 ##########
 User guide
 ##########
 
-.. .. toctree::
-..    :maxdepth: 2
-.. .. :titlesonly:
+Python consumers of data and service discovery information should use the Repertoire client, `rubin.repertoire`.
+This library is available from PyPI and can be declared as a dependency or installed with pip in the normal way:
+
+.. prompt:: bash
+
+   pip install rubin-repertoire
+
+Consumers in other languages, such as JavaScript, can use the Repertoire server API directly.
+
+.. toctree::
+   :caption: Python client
+
+   initialization
+   services
+   datasets
+   applications
+
+.. toctree::
+   :caption: Server
+
+   server-api
