+++ docs/user-guide/server-api.rst
@@ -1,0 +1,5 @@
+#####################
+Repertoire server API
+#####################
+
+This is a stub page for the API.
